requirements.txt
requests
trimesh
numpy
PyGithub
